~Auto: Commit

+++ Tk/System.php
@@ -188,9 +188,9 @@ class System
     {
         $released = new \DateTime();
         if (is_file(Config::makePath('/version'))) {
-            $released = Date::create(filemtime(Config::makePath('/version')));
+            $released = Date::create((int)filemtime(Config::makePath('/version')));
         } else if (is_file(Config::makePath('/version.md'))) {
-            $released = Date::create(filemtime(Config::makePath('/version.md')));
+            $released = Date::create((int)filemtime(Config::makePath('/version.md')));
         }
         return $released;
     }
